refactor: extract game state into hook
