+++ src/App.jsx
@@ -1,54 +1,26 @@
-import React, { useState } from 'react';
+import React from 'react';
 import Board from './components/Board';
 import DiceRoll from './components/DiceRoll';
 import Controls from './components/Controls';
 import TailwindOptions from './components/TailwindOptions';
+import useGameState from './hooks/useGameState';
 
-// The main application keeps all game state and passes data down to
-// presentational components. Child components communicate actions
-// through callback props which allows App to remain the single source
-// of truth.
+// The main application delegates game state to the custom useGameState
+// hook while remaining the central coordinator for rendering and event
+// wiring between child components.
 export default function App() {
-  const [dice, setDice] = useState([1, 1, 1]);
-  const [selectedPair, setSelectedPair] = useState(null);
-  const [pieces, setPieces] = useState([]); // {id, r, step}
-  const [lanes] = useState([
-    { sum: 2, L: 3 },
-    { sum: 3, L: 4 },
-    { sum: 4, L: 5 },
-    { sum: 5, L: 6 },
-    { sum: 6, L: 7 },
-    { sum: 7, L: 8 },
-    { sum: 8, L: 7 },
-    { sum: 9, L: 6 },
-    { sum: 10, L: 5 },
-    { sum: 11, L: 4 },
-    { sum: 12, L: 3 },
-  ]);
-
-  const rollDice = () => {
-    const next = Array.from({ length: 3 }, () => Math.floor(Math.random() * 6) + 1);
-    setDice(next);
-    setSelectedPair(null);
-  };
-
-  const handleSelectPair = pair => {
-    setSelectedPair(pair);
-  };
-
-  const handleMove = () => {
-    if (!selectedPair) return;
-    const sum = selectedPair[0] + selectedPair[1];
-    const laneIndex = lanes.findIndex(l => l.sum === sum);
-    if (laneIndex === -1) return;
-    const existing = pieces.find(p => p.r === laneIndex);
-    if (existing) {
-      setPieces(pieces.map(p => p.r === laneIndex ? { ...p, step: p.step + 1 } : p));
-    } else {
-      setPieces([...pieces, { id: Date.now(), r: laneIndex, step: 1 }]);
-    }
-    setSelectedPair(null);
-  };
+  const {
+    dice,
+    selectedPair,
+    pieces,
+    lanes,
+    rollDice,
+    selectPair,
+    move,
+    swoop,
+    bank,
+    resetPieces,
+  } = useGameState();
 
   const handleCellClick = (r, step) => {
     // Placeholder interaction for board clicks
@@ -57,16 +29,18 @@ export default function App() {
 
   return (
     <div className="p-4 space-y-4">
-      <TailwindOptions onReset={() => setPieces([])} />
+      <TailwindOptions onReset={resetPieces} />
       <Controls
         onRoll={rollDice}
-        onMove={handleMove}
+        onMove={move}
+        onSwoop={swoop}
+        onBank={bank}
         moveDisabled={!selectedPair}
       />
       <DiceRoll
         dice={dice}
         selectedPair={selectedPair}
-        onSelectPair={handleSelectPair}
+        onSelectPair={selectPair}
       />
       <Board
         lanes={lanes}
